fix(autoware_auto_geometry): move headers to a separate directory
