+++ common/autoware_auto_geometry/include/autoware_auto_geometry/bounding_box/rotating_calipers.hpp
@@ -19,9 +19,9 @@
 
 #ifndef GEOMETRY__BOUNDING_BOX__ROTATING_CALIPERS_HPP_
 #define GEOMETRY__BOUNDING_BOX__ROTATING_CALIPERS_HPP_
-#include <geometry/bounding_box/bounding_box_common.hpp>
-#include <geometry/common_2d.hpp>
-#include <geometry/convex_hull.hpp>
+#include "autoware_auto_geometry/bounding_box/bounding_box_common.hpp>
+#include "autoware_auto_geometry/common_2d.hpp>
+#include "autoware_auto_geometry/convex_hull.hpp>
 
 #include <algorithm>
 #include <cstring>
