+++ common/autoware_auto_geometry/include/autoware_auto_geometry/lookup_table.hpp
@@ -21,7 +21,7 @@
 #define GEOMETRY__LOOKUP_TABLE_HPP_
 
 #include "common/types.hpp"
-#include "geometry/interval.hpp"
+#include "autoware_auto_geometry/interval.hpp"
 
 #include <cmath>
 #include <stdexcept>
