+++ common/autoware_auto_geometry/include/autoware_auto_geometry/spatial_hash.hpp
@@ -21,8 +21,8 @@
 #define GEOMETRY__SPATIAL_HASH_HPP_
 
 #include <common/types.hpp>
-#include <geometry/spatial_hash_config.hpp>
-#include <geometry/visibility_control.hpp>
+#include "autoware_auto_geometry/spatial_hash_config.hpp>
+#include "autoware_auto_geometry/visibility_control.hpp>
 
 #include <unordered_map>
 #include <utility>
