+++ common/autoware_auto_geometry/include/autoware_auto_geometry/visibility_control.hpp
