+++ common/autoware_auto_geometry/src/bounding_box.cpp
@@ -15,12 +15,12 @@
 // Co-developed by Tier IV, Inc. and Apex.AI, Inc.
 
 #include <autoware_auto_tf2/tf2_autoware_auto_msgs.hpp>
-#include <geometry/bounding_box/bounding_box_common.hpp>
-#include <geometry/bounding_box/eigenbox_2d.hpp>
+#include "autoware_auto_geometry/bounding_box/bounding_box_common.hpp>
+#include "autoware_auto_geometry/bounding_box/eigenbox_2d.hpp>
 // cspell: ignore eigenbox
-#include <geometry/bounding_box/lfit.hpp>
+#include "autoware_auto_geometry/bounding_box/lfit.hpp>
 // cspell: ignore lfit
-#include <geometry/bounding_box/rotating_calipers.hpp>
+#include "autoware_auto_geometry/bounding_box/rotating_calipers.hpp>
 
 #include <geometry_msgs/msg/point32.hpp>
