+++ common/autoware_auto_geometry/test/src/test_common_2d.cpp
@@ -14,7 +14,7 @@
 //
 // Co-developed by Tier IV, Inc. and Apex.AI, Inc.
 
-#include <geometry/common_2d.hpp>
+#include "autoware_auto_geometry/common_2d.hpp>
 
 #include <autoware_auto_perception_msgs/msg/point_clusters.hpp>
 #include <geometry_msgs/msg/point32.hpp>
