+++ common/autoware_auto_geometry/test/src/test_intersection.cpp
@@ -14,8 +14,8 @@
 
 // Co-developed by Tier IV, Inc. and Apex.AI, Inc.
 
-#include <geometry/convex_hull.hpp>
-#include <geometry/intersection.hpp>
+#include "autoware_auto_geometry/convex_hull.hpp>
+#include "autoware_auto_geometry/intersection.hpp>
 
 #include <gtest/gtest.h>
